working nodejs test

+++ include/Volume.h
@@ -33,23 +33,23 @@ class CVolumeMetadata {
 private:
 
   struct CSegments {
-    size_t                          count;
-    std::vector<size_t>             sizes;
-    std::vector<vmml::AABB<size_t>> boundsVolume;
-    std::vector<vmml::AABB<size_t>> boundsWorld;
+    int64_t                          count;
+    std::vector<int64_t>             sizes;
+    std::vector<vmml::AABB<int64_t>> boundsVolume;
+    std::vector<vmml::AABB<int64_t>> boundsWorld;
 
     CSegments(const CVolumeMetadata &meta, const std::vector<unsigned char> &raw_bboxes, const std::vector<unsigned char> &raw_sizes);
   };
 
-  vmml::AABB<size_t>         physical_offset;
-  vmml::Vector<3, size_t>    volume_dimensions;
-  vmml::Vector<3, size_t>    voxel_resolution;
+  vmml::AABB<int64_t>        physical_offset;
+  vmml::Vector<3, int64_t>   volume_dimensions;
+  vmml::Vector<3, int64_t>   voxel_resolution;
   std::string                resolution_units;
   MetaDataType               segment_id_type;
   MetaDataType               segment_bbox_type;
   MetaDataType               segment_size_type;
   uint8_t                    segment_id_type_size;
-  size_t                     segment_count;
+  int64_t                    segment_count;
   CSegments                * segments;
 
 public:
@@ -57,12 +57,12 @@ public:
   CVolumeMetadata(const std::vector<unsigned char> &raw_json, const std::vector<unsigned char> &raw_bboxes, const std::vector<unsigned char> &raw_sizes);
   ~CVolumeMetadata();
 
-  const vmml::AABB<size_t> &      GetPhysicalBounds() const;
-  const vmml::Vector<3, size_t> & GetVolumeDimensions() const;
-  const vmml::Vector<3, size_t> & GetVoxelResolution() const;
-  const std::string &             GetResolutionUnit() const;
-  uint8_t                         GetSegmentTypeSize() const;
-  size_t                          GetSegmentCount() const;
+  const vmml::AABB<int64_t> &      GetPhysicalBounds() const;
+  const vmml::Vector<3, int64_t> & GetVolumeDimensions() const;
+  const vmml::Vector<3, int64_t> & GetVoxelResolution() const;
+  const std::string &              GetResolutionUnit() const;
+  uint8_t                          GetSegmentTypeSize() const;
+  int64_t                          GetSegmentCount() const;
 
 };
 
@@ -70,13 +70,13 @@ public:
 
 class CSegmentation {
 protected:
-  const vmml::Vector<3, size_t> dimensions_;
+  const vmml::Vector<3, int64_t> dimensions_;
 
-  CSegmentation(const vmml::Vector<3, size_t> & dimensions);
+  CSegmentation(const vmml::Vector<3, int64_t> & dimensions);
 
 public:
-  virtual uint32_t operator()(size_t x, size_t y, size_t z) const;
-  virtual uint32_t operator()(const vmml::Vector<3, size_t> & pos) const;
+  virtual uint32_t operator()(int64_t x, int64_t y, int64_t z) const;
+  virtual uint32_t operator()(const vmml::Vector<3, int64_t> & pos) const;
 
 };
 
@@ -86,9 +86,9 @@ class CSegmentationUChar : public CSegmentation {
 private:
   const uint8_t * segmentation_;
 public:
-  CSegmentationUChar(const vmml::Vector<3, size_t> &dimensions, const uint8_t * segmentation);
-  uint32_t operator()(size_t x, size_t y, size_t z) const override;
-  uint32_t operator()(const vmml::Vector<3, size_t> & pos) const override;
+  CSegmentationUChar(const vmml::Vector<3, int64_t> &dimensions, const uint8_t * segmentation);
+  uint32_t operator()(int64_t x, int64_t y, int64_t z) const override;
+  uint32_t operator()(const vmml::Vector<3, int64_t> & pos) const override;
 };
 
 /*****************************************************************/
@@ -97,9 +97,9 @@ class CSegmentationUShort : public CSegmentation {
 private:
   const uint16_t * segmentation_;
 public:
-  CSegmentationUShort(const vmml::Vector<3, size_t> &dimensions, const uint16_t * segmentation);
-  uint32_t operator()(size_t x, size_t y, size_t z) const override;
-  uint32_t operator()(const vmml::Vector<3, size_t> & pos) const override;
+  CSegmentationUShort(const vmml::Vector<3, int64_t> &dimensions, const uint16_t * segmentation);
+  uint32_t operator()(int64_t x, int64_t y, int64_t z) const override;
+  uint32_t operator()(const vmml::Vector<3, int64_t> & pos) const override;
 };
 
 /*****************************************************************/
@@ -108,9 +108,9 @@ class CSegmentationUInt : public CSegmentation {
 private:
   const uint32_t * segmentation_;
 public:
-  CSegmentationUInt(const vmml::Vector<3, size_t> &dimensions, const uint32_t * segmentation);
-  uint32_t operator()(size_t x, size_t y, size_t z) const override;
-  uint32_t operator()(const vmml::Vector<3, size_t> & pos) const override;
+  CSegmentationUInt(const vmml::Vector<3, int64_t> &dimensions, const uint32_t * segmentation);
+  uint32_t operator()(int64_t x, int64_t y, int64_t z) const override;
+  uint32_t operator()(const vmml::Vector<3, int64_t> & pos) const override;
 };
 
 /*****************************************************************/
@@ -118,7 +118,7 @@ public:
 class CVolume {
   private:
 
-  static vmml::AABB<size_t>        empty_bbox;
+  static vmml::AABB<int64_t>       empty_bbox;
 
   std::unique_ptr<CVolumeMetadata> meta_;
 
@@ -129,13 +129,13 @@ class CVolume {
   CVolume(std::unique_ptr<CVolumeMetadata> &&meta, const std::vector<unsigned char> &raw_segmentation);
   ~CVolume();
 
-  const vmml::AABB<size_t> &      GetPhysicalBounds() const;
-  const vmml::Vector<3, size_t> & GetVoxelResolution() const;
-  size_t                          GetSegmentCount() const;
-  const vmml::AABB<size_t> &      GetSegmentBoundsWorld(size_t segID) const;
-  const vmml::AABB<size_t> &      GetSegmentBoundsVolume(size_t segID) const;
-  size_t                          GetSegmentSizeVoxel(size_t segID) const;
-  const CSegmentation *           GetSegmentation() const;
+  const vmml::AABB<int64_t> &      GetPhysicalBounds() const;
+  const vmml::Vector<3, int64_t> & GetVoxelResolution() const;
+  int64_t                          GetSegmentCount() const;
+  const vmml::AABB<int64_t> &      GetSegmentBoundsWorld(int64_t segID) const;
+  const vmml::AABB<int64_t> &      GetSegmentBoundsVolume(int64_t segID) const;
+  int64_t                          GetSegmentSizeVoxel(int64_t segID) const;
+  const CSegmentation *            GetSegmentation() const;
 
 
 };
